Move classes representing GCP based transformations from app to analysis, and make ready for exposure to public API
